Added the ability to upload thumbnail images

+++ app/Http/Controllers/FullApiServiceController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Modules\Core\Services\ImageService;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 abstract class FullApiServiceController extends ApiServiceController
 {
+    protected ImageService $imageService;
 
     public function create(Request $request): JsonResponse
     {
@@ -31,6 +33,11 @@ abstract class FullApiServiceController extends ApiServiceController
 
     public function delete(int $id)
     {
+        $this->imageService->deleteIfExists($id);
+        if ($this->service->hasErrors()) {
+            return $this->createErrorResponse();
+        }
+
         $this->service->delete($id);
         if ($this->service->hasErrors()) {
             return $this->createErrorResponse();
